Expanded WildernessSpawner
Updated the rolls for the Wilderness Spawner, to use a more easily
expandable/modular process to roll for tiles. Also added MineableObjects
and the WarpPad to the map generation, with the Player being spawned at
the location of the WarpPad.

Also added/updated some documentation and comments and capitalized
public prefab fields for style consistency.