Code & Documentation clean-up
Refactored the map randomization and instantiation into a public
CreateMap() function, which contains the following private functions in
sequence:
- randomizeMap() to set the size and types of map used
- setPlayerStart() to initialize the player and warp pad position
- instantiateMap() to actually instantiate the objects of the map.

Furthermore, the instantiation of creating a warp pad & player and
creating a new mine into the placePlayer() and placeMine() functions.

Finally, cleaned up the documentation for the WildernessSpawner for
consistency and completeness.